Decoupling : Final stage
- All the constants imported from the definitions file
- MrbPicWorker defined in worker.py
- Removed typing dependency (unavailable on our machines)
- Moved lens undistorion functions from init and undistort.py to lens.py
