+++ octoprint_mrbeam/camera/__init__.py
@@ -8,48 +8,18 @@ from abc import ABCMeta, abstractmethod
 import os
 from octoprint_mrbeam.mrb_logger import mrb_logger
 from octoprint_mrbeam.util.img import differed_imwrite
+from .definitions import SUCCESS_WRITE_RETVAL, N, S, W, E, QD_KEYS, RATIO_W, RATIO_H, OFFSET_W, OFFSET_H, DIFF_TOLERANCE
 # Python 3 : use ABC instead of ABCMeta
-
-SUCCESS_WRITE_RETVAL = 1
 
 from octoprint_mrbeam.util.log import logtime, logme
 
 try:
 	import picamera
 	PICAMERA_AVAILABLE = True
-except ImportError as e:
+except (ImportError, OSError) as e:
 	PICAMERA_AVAILABLE = False
 	logging.getLogger("octoprint.plugins.mrbeam.util.camera").error(
 		"Could not import module 'mrbcamera'. Disabling camera integration. (%s: %s)", e.__class__.__name__, e)
-
-
-RESOLUTIONS = {
-		'1000x780':  (1000, 780),
-		'1920x1080': (1920, 1080),
-		'2000x1440': (2000, 1440),
-		'2048x1536': (2048, 1536),
-		'2592x1944': (2592, 1944),
-		'2592x1952': (2592, 1952),
-}
-
-N, W, S, E = 'N','W','S','E'
-NW,NE,SW,SE = N+W, N+E, S+W, S+E
-QD_KEYS = [NW,NE,SW,SE]
-
-# Size of the corner search area
-RATIO_W, RATIO_H = Fraction(1, 6), Fraction(1, 4)
-# Padding distance from the edges of the image (The markers are never pressed against the border)
-OFFSET_W, OFFSET_H = Fraction(0, 36), Fraction(0, 20)
-
-LEGACY_STILL_RES = RESOLUTIONS['2048x1536'] # from octoprint_mrbeam __init___ : get_settings_defaults
-DEFAULT_STILL_RES = RESOLUTIONS['2592x1944']  # Be careful : Resolutions accepted as increments of 32 horizontally and 16 vertically
-
-# threshold; 2 consecutive pictures need to have a minimum difference
-# before being undistorted and served
-DIFF_TOLERANCE = 50
-
-TARGET_AVG_ROI_BRIGHTNESS = 170
-BRIGHTNESS_TOLERANCE = 40 # TODO Keep the brightness of the images tolerable
 
 class Camera:
 	__metaclass__ = ABCMeta
@@ -63,65 +33,6 @@ class Camera:
 		pass
 
 	pass
-
-class MrbPicWorker(object):
-	"""
-	The class that take care of buffering the pictures taken from the camera.
-	It can also do frame-by-frame work, but it could hurt recording speed (better
-	to split the work on a different thread)
-	See "Advanced Recipies" in the PiCamera tutorials:
-	https://picamera.readthedocs.io/en/release-1.13/recipes2.html
-	"""
-	def __init__(self, maxSize=3, debug=False):
-		self.images = []
-		self.firstImg = True
-		self.buffers = [io.BytesIO() for _ in range(maxSize)]
-		self.bufferIndex = 0
-		self.latest = None
-		self.avg_roi_brightness = {}
-		assert(maxSize > 0)
-		self._maxSize = maxSize
-		self.times = []  # exposure time values
-		self.busy = Event()
-		self._logger = mrb_logger("mrbeam.camera.MrbPicWorker")
-		if debug: self._logger.setLevel(logging.DEBUG)
-		else: self._logger.setLevel(logging.WARNING)
-
-	def currentBuf(self):
-		return self.buffers[self.bufferIndex]
-
-	def flush(self):
-		# Is called when the camera is done writing the whole image into the buffer
-		self.currentBuf().seek(0)
-		self.latest = cv2.imdecode(np.fromstring(self.currentBuf().getvalue(), np.int8),
-								   cv2.IMREAD_COLOR)
-		self.bufferIndex = (self.bufferIndex + 1) % self._maxSize
-		self.currentBuf().seek(0)
-		self.currentBuf().truncate()
-		self.avg_roi_brightness = brightness_result(self.latest)
-		# TODO adjust camera shutter speed with these brightness measurements
-		self.busy.clear()
-
-	def write(self, buf):  # (self, buf: bytearray):
-		"""
-		Write into the current buffer.
-		Will automatically change buffer when a new JPEG image is detected.
-		"""
-		if buf.startswith(b'\xff\xd8') and self.currentBuf().tell() > 0:
-			# New frame; and the current buffer is not flushed.
-			self.flush()
-		# Add the buffer to the currently selected buffer
-		self.busy.set()
-		self.currentBuf().write(buf)
-
-	def saveImg(self, path, n=1):
-		"""Saves the last image or the n-th last buffer"""
-		# Unused atm
-		assert(0 < n <= self._maxSize)
-		f = io.open(path, 'wb')
-		ret = f.write(self.buffers[-n])
-		f.close()
-		return ret
 
 def brightness_result(pic):
 	"""
